Reorganize code
Reorganize the code between analyze, compile and utils.
  - analyze: Parse the .abc file, convert lines of .abc code in tokens
(Operations) and get meta informations
  - compile: Convert operations and meta informations to bytecode
  - utils: Helpers and miscellanous function used through the program
